@@ -23,14 +23,15 @@ import org.mybatis.dynamic.sql.select.QueryExpressionDSL;
 import org.mybatis.dynamic.sql.select.SelectModel;
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.sql.Date;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static org.mybatis.dynamic.sql.SqlBuilder.isEqualTo;
-import static org.mybatis.dynamic.sql.SqlBuilder.select;
+import static org.mybatis.dynamic.sql.SqlBuilder.*;
 
 @Service
 @Slf4j
@@ -330,21 +331,13 @@ public class CommentService {
     public PageInfo<CommentAdminVo> queryFirstCommentAdminList(CommentListAdminQueryDTO dto) {
         Integer page = dto.getPage();
         Integer size = dto.getSize();
-        Integer newsId = dto.getNewsId();
+        final String content = StringUtils.hasLength(dto.getContent()) ? dto.getContent() + "%" : null;
+//        final String name = StringUtils.hasLength(dto.getName()) ? dto.getName() + "%" : null;
+        final String title = StringUtils.hasLength(dto.getTitle()) ? dto.getTitle() + "%" : null;
 
-        QueryExpressionDSL<SelectModel>.QueryExpressionWhereBuilder builder = select(FirstCommentMapper.selectList)
+        SelectStatementProvider firstStatement = select(FirstCommentMapper.selectList)
                 .from(FirstCommentDynamicSqlSupport.firstComment)
-                .where();
-
-        if (newsId != null) {
-            // 数据库不存在该 newsId
-            if (newsMapper.count(c -> c.where(NewsDynamicSqlSupport.newsId, isEqualTo(newsId))) != 1) {
-                throw new AlertException(400, "该新闻不存在");
-            }
-            builder.and(FirstCommentDynamicSqlSupport.newsId, isEqualTo(newsId));
-        }
-
-        SelectStatementProvider firstStatement = builder
+                .where(FirstCommentDynamicSqlSupport.content,isLikeWhenPresent(content))
                 .build()
                 .render(RenderingStrategies.MYBATIS3);
 
@@ -358,20 +351,20 @@ public class CommentService {
             vo.setPublishTime(item.getPublishTime().getTime());
             vo.setContent(item.getContent());
             vo.setLikeNumber(item.getLikeNumber());
-            Optional<News> news = newsMapper.selectOne(s -> s.where(NewsDynamicSqlSupport.newsId, isEqualTo(item.getNewsId())));
+            Optional<News> news = newsMapper.selectOne(s -> s.where(NewsDynamicSqlSupport.newsId, isEqualTo(item.getNewsId()))
+                    .and(NewsDynamicSqlSupport.title,isLikeWhenPresent(title)));
             if (news.isPresent()) {
                 vo.setNewsContent(news.get().getContent());
                 vo.setNewsTitle(news.get().getTitle());
             } else {
-                vo.setNewsContent(null);
-                vo.setNewsTitle(null);
+                return null;
             }
             vo.setPublisher(queryUser(item.getPublisherId()));
             vo.setParentCommentContent(null);
             vo.setParentCommentId(-1);
             return vo;
-        }).collect(Collectors.toList());
-
+        }).filter(Objects::nonNull).collect(Collectors.toList());
+        queryPage.setTotal(result.size());
         return generatePageInfo(queryPage, result);
     }
 
@@ -383,6 +376,9 @@ public class CommentService {
     public PageInfo<CommentAdminVo> querySecondCommentAdminList(CommentListAdminQueryDTO dto) {
         Integer page = dto.getPage();
         Integer size = dto.getSize();
+        final String content = StringUtils.hasLength(dto.getContent()) ? dto.getContent() + "%" : null;
+//        final String name = StringUtils.hasLength(dto.getName()) ? dto.getName() + "%" : null;
+        final String title = StringUtils.hasLength(dto.getTitle()) ? dto.getTitle() + "%" : null;
 //        Integer newsId = dto.getNewsId();
 
 //        QueryExpressionDSL<SelectModel>.QueryExpressionWhereBuilder builder = select(FirstCommentMapper.selectList)
@@ -403,7 +399,7 @@ public class CommentService {
 
         SelectStatementProvider secondStatement = select(SecondCommentMapper.selectList)
                 .from(SecondCommentDynamicSqlSupport.secondComment)
-                .where()
+                .where(SecondCommentDynamicSqlSupport.content,isLikeWhenPresent(content))
                 .build()
                 .render(RenderingStrategies.MYBATIS3);
 
@@ -418,13 +414,13 @@ public class CommentService {
                 vo.setParentCommentId(firstComment.get().getCommentId());
                 vo.setParentCommentContent(firstComment.get().getContent());
 
-                Optional<News> news = newsMapper.selectOne(s -> s.where(NewsDynamicSqlSupport.newsId, isEqualTo(firstComment.get().getNewsId())));
+                Optional<News> news = newsMapper.selectOne(s -> s.where(NewsDynamicSqlSupport.newsId, isEqualTo(firstComment.get().getNewsId()))
+                        .and(NewsDynamicSqlSupport.title,isLikeWhenPresent(title)));
                 if (news.isPresent()) {
                     vo.setNewsContent(news.get().getContent());
                     vo.setNewsTitle(news.get().getTitle());
                 } else {
-                    vo.setNewsContent(null);
-                    vo.setNewsTitle(null);
+                    return null;
                 }
             }
             vo.setPublisher(queryUser(item.getPublisherId()));
@@ -435,8 +431,9 @@ public class CommentService {
             vo.setLikeNumber(item.getLikeNumber());
 
             return vo;
-        }).collect(Collectors.toList());
+        }).filter(Objects::nonNull).collect(Collectors.toList());
 
+        queryPage.setTotal(result.size());
         return generatePageInfo(queryPage, result);
     }
 
