@@ -1,14 +1,17 @@
-package com.news.NS.domain.dto;
+package com.news.NS.domain.dto.Complaint;
 
 import lombok.Data;
 import lombok.ToString;
-
+import javax.validation.constraints.NotNull;
 import java.sql.Timestamp;
 
 @Data
 @ToString
 public class ComplaintCreateDTO {
+    @NotNull
     private Integer newsId;
+    @NotNull
     private Integer complainerId;
+    @NotNull
     private String reason;
 }
