@@ -1,11 +1,15 @@
-package com.news.NS.domain.dto;
+package com.news.NS.domain.dto.Complaint;
 
 import lombok.Data;
 import lombok.ToString;
 
+import javax.validation.constraints.NotNull;
+
 @Data
 @ToString
 public class ComplaintDeleteDTO {
+    @NotNull
     private Integer newsId;
+    @NotNull
     private Integer complaintId;
 }
