@@ -3,26 +3,32 @@ package com.news.NS.service;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.news.NS.common.AlertException;
-import com.news.NS.common.CommonConstant;
 import com.news.NS.common.domain.PageInfo;
 import com.news.NS.common.domain.ResultCode;
 import com.news.NS.domain.Complaint;
+import com.news.NS.domain.dto.Complaint.ComplaintCreateDTO;
+import com.news.NS.domain.dto.Complaint.ComplaintDeleteDTO;
+import com.news.NS.domain.dto.Complaint.ComplaintModifyDTO;
+import com.news.NS.domain.dto.Complaint.ComplaintSearchDTO;
 import com.news.NS.domain.News;
 import com.news.NS.domain.User;
-import com.news.NS.domain.dto.ComplaintCreateDTO;
-import com.news.NS.domain.dto.ComplaintDeleteDTO;
-import com.news.NS.domain.dto.ComplaintModifyDTO;
-import com.news.NS.domain.dto.ComplaintSearchDTO;
+import com.news.NS.domain.dto.*;
+import com.news.NS.domain.vo.ComplaintListVo;
+import com.news.NS.domain.vo.NewsListVo;
 import com.news.NS.mapper.*;
 import org.mybatis.dynamic.sql.render.RenderingStrategies;
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider;
 import org.mybatis.dynamic.sql.update.render.UpdateStatementProvider;
+import org.springframework.cglib.beans.BeanCopier;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static org.mybatis.dynamic.sql.SqlBuilder.*;
 
@@ -41,17 +47,23 @@ public class ComplaintService {
     public void addNewComplaint(ComplaintCreateDTO complaintCreateDTO) {
         Complaint temp = new Complaint();
         //检查用户是否存在。
-        if(!userMapper.selectByPrimaryKey(complaintCreateDTO.getComplainerId()).isPresent()){
+        if(userMapper.count(c->c.where(UserDynamicSqlSupport.userId,isEqualTo(complaintCreateDTO.getComplainerId()))) == 0){
             throw new AlertException(ResultCode.USER_NOT_EXIST);
         }
         temp.setComplainerId(complaintCreateDTO.getComplainerId());
         //检查新闻是否存在
-        if(!newsMapper.selectByPrimaryKey(complaintCreateDTO.getNewsId()).isPresent()){
-            throw new AlertException(500,"id为"+complaintCreateDTO.getNewsId()+"新闻不存在");
+        if(newsMapper.count(c->c.where(NewsDynamicSqlSupport.newsId,isEqualTo(complaintCreateDTO.getNewsId()))) == 0){
+            throw new AlertException(500,"新闻不存在");
         }
         temp.setNewsId(complaintCreateDTO.getNewsId());
+        String st = complaintCreateDTO.getReason();
+        if(StringUtils.hasLength(st)){
+            temp.setComplaintReason(complaintCreateDTO.getReason());
+        } else {
+            throw new AlertException(ResultCode.PARAM_IS_BLANK);
+        }
+
         temp.setComplaintTime(Timestamp.valueOf(LocalDateTime.now()));
-        temp.setComplaintReason(complaintCreateDTO.getReason());
         complaintMapper.insert(temp);
     }
 
@@ -62,6 +74,9 @@ public class ComplaintService {
     }
 
     public void modifyComplaint(ComplaintModifyDTO dto) {
+        if(!StringUtils.hasLength(dto.getReason())){
+            throw new AlertException(ResultCode.PARAM_IS_BLANK);
+        }
         UpdateStatementProvider updateStatement = update(ComplaintDynamicSqlSupport.complaint)
                 .set(ComplaintDynamicSqlSupport.complaintReason).equalTo(dto.getReason())
                 .where(ComplaintDynamicSqlSupport.complainerId,isEqualTo(dto.getComplaintId()))
@@ -71,19 +86,48 @@ public class ComplaintService {
         }
     }
 
-    public PageInfo<Complaint> searchComplaint(ComplaintSearchDTO<String> dto) {
-        String likeKeyword = "%" + dto.getSearchKeyword() + "%";
+    public PageInfo<ComplaintListVo> searchComplaint(ComplaintListDTO dto) {
+        final String reason = StringUtils.hasLength(dto.getReason()) ? dto.getReason() + "%" : null;
+        final String name = StringUtils.hasLength(dto.getName()) ? dto.getName() + "%" : null;
+        final String title = StringUtils.hasLength(dto.getTitle()) ? dto.getTitle() + "%" : null;
+
 
         SelectStatementProvider selectStatement = select(complaintMapper.selectList)
                 .from(ComplaintDynamicSqlSupport.complaint)
-                .where(ComplaintDynamicSqlSupport.complaintReason,isLike(likeKeyword))
+                .where(ComplaintDynamicSqlSupport.complaintReason,isLike(reason))
                 .build().render(RenderingStrategies.MYBATIS3);
 
         Page<Complaint> queryPageData = PageHelper.startPage(dto.getPage(), dto.getSize());
         List<Complaint> complaints = complaintMapper.selectMany(selectStatement);
+        BeanCopier copier = BeanCopier.create(Complaint.class, ComplaintListVo.class, false);
 
-        return packing(complaints,dto.getPage(), queryPageData.getTotal());
+        List<ComplaintListVo> complaintListVos = complaints.stream().map(item ->{
+            ComplaintListVo complaintListVo = new ComplaintListVo();
+            copier.copy(item,complaintListVo,null);
+            Optional<News> news = newsMapper.selectOne(s -> s.where(NewsDynamicSqlSupport.newsId, isEqualTo(item.getNewsId()))
+                    .and(NewsDynamicSqlSupport.title,isLikeWhenPresent(title)));
+            if (news.isPresent()) {
+                complaintListVo.setNewsContent(news.get().getContent());
+                complaintListVo.setNewsTitle(news.get().getTitle());
+            } else {
+                return null;
+            }
+            Optional<User> user = userMapper.selectOne(s -> s.where(UserDynamicSqlSupport.userId, isEqualTo(item.getComplainerId()))
+                    .and(UserDynamicSqlSupport.username,isLikeWhenPresent(name)));
+            if (user.isPresent()) {
+                complaintListVo.setComplainerName(user.get().getUsername());
+            } else {
+                return null;
+            }
+            return complaintListVo;
+        }).filter(Objects::nonNull).collect(Collectors.toList());
+        PageInfo<ComplaintListVo> pageInfo = new PageInfo<>();
+        pageInfo.setPageData(complaintListVos);
+        pageInfo.setPage(dto.getPage());
+        pageInfo.setTotalSize(queryPageData.getTotal());
+        return pageInfo;
     }
+
     private PageInfo<Complaint> packing(List<Complaint> complaints,Integer page,long total){
         /*
          * 把多个News分页封装成Pageinfo然后返回
@@ -96,5 +140,17 @@ public class ComplaintService {
         pageInfo.setPage(page);
         pageInfo.setTotalSize(total);
         return pageInfo;
+    }
+
+    public PageInfo<Complaint> getByComplainerId(ComplaintSearchDTO<Integer> dto) {
+        SelectStatementProvider selectStatement = select(complaintMapper.selectList)
+                .from(ComplaintDynamicSqlSupport.complaint)
+                .where(ComplaintDynamicSqlSupport.complainerId,isEqualTo(dto.getSearchKeyword()))
+                .build().render(RenderingStrategies.MYBATIS3);
+
+        Page<Complaint> queryPageData = PageHelper.startPage(dto.getPage(), dto.getSize());
+        List<Complaint> complaints = complaintMapper.selectMany(selectStatement);
+
+        return packing(complaints,dto.getPage(), queryPageData.getTotal());
     }
 }
